# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Phlote API
  description: API documentation for Phlote application
  version: 1.0.0
servers:
  - url: /api
    description: Default API server
...
paths:
...
  /applications:
    get:
      tags:
        - Applications
      summary: Get all applications
      description: Retrieve all applications
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /creator-applications:
    post:
      tags:
        - Creator Applications
      summary: Create a creator application
      description: Submit a new creator application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatorApplication"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  applicationId:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
    EthereumAddress:
      type: string
      pattern: "^0x[a-fA-F0-9]{40}$"
      description: Ethereum address (40 hex characters)
      example: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
...
    Application:
      type: object
      description: Creator application document
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        city:
          type: string
        info:
          type: string
        workLink:
          type: string
        ethAddress:
          $ref: "#/components/schemas/EthereumAddress"
        tracks:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              id:
                type: string
    CreatorApplication:
      type: object
      required:
        - firstName
        - lastName
        - email
        - city
        - tracks
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        city:
          type: string
          minLength: 1
        info:
          type: string
        workLink:
          type: string
        ethAddress:
          $ref: "#/components/schemas/EthereumAddress"
        tracks:
          type: array
          items:
            type: object
            required:
              - name
              - id
            properties:
              name:
                type: string
                minLength: 1
              id:
                type: string
                minLength: 1
...
  responses:
    BadRequest:
      description: Bad request - Invalid parameters or request body
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid request
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Internal server error